LPS-99657 Move to 'internal'
