+++ modules/apps/portal-security/portal-security-auto-login/src/main/java/com/liferay/portal/security/auto/login/internal/request/header/configuration/definition/RequestHeaderAutoLoginCompanyServiceConfigurationPidMapping.java
@@ -12,29 +12,29 @@
  * details.
  */
 
-package com.liferay.portal.security.auto.login.basic.auth.header.configuration.definition;
+package com.liferay.portal.security.auto.login.internal.request.header.configuration.definition;
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
-import com.liferay.portal.security.auto.login.basic.auth.header.configuration.BasicAuthHeaderAutoLoginConfiguration;
-import com.liferay.portal.security.auto.login.internal.basic.auth.header.constants.BasicAuthHeaderAutoLoginConstants;
+import com.liferay.portal.security.auto.login.internal.request.header.configuration.RequestHeaderAutoLoginConfiguration;
+import com.liferay.portal.security.auto.login.internal.request.header.constants.RequestHeaderAutoLoginConstants;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Tomas Polesovsky
+ * @author Mika Koivisto
  */
 @Component(service = ConfigurationPidMapping.class)
-public class BasicAuthHeaderAutoLoginCompanyServiceConfigurationPidMapping
+public class RequestHeaderAutoLoginCompanyServiceConfigurationPidMapping
 	implements ConfigurationPidMapping {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
-		return BasicAuthHeaderAutoLoginConfiguration.class;
+		return RequestHeaderAutoLoginConfiguration.class;
 	}
 
 	@Override
 	public String getConfigurationPid() {
-		return BasicAuthHeaderAutoLoginConstants.SERVICE_NAME;
+		return RequestHeaderAutoLoginConstants.SERVICE_NAME;
 	}
 
 }
